refactor: Update lazy loading for images in dashboard and sequence pages
